api/src/requirements.txt
boto3==1.17.98
celery[redis]==5.1.2
classutilities==0.2.0
Django==3.2.4
psycopg2-binary==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.7.0
flake8==3.9.2
flake8-annotations==2.6.2
flake8-bugbear==21.4.3
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
flake8-type-annotations==0.1.0
darglint==1.8.0
black==21.7b0